kreiran authMiddleware.js fajl
-nastavak rada na projektu Book Store App
-kreiran folder middleware i authMiddleware.js fajl
-u fajlove bookRoutes i userRoutes.js fajl dodana funkcija authMiddleware
-manje izmjene u postojecim fajlovima

+++ final-project/controllers/userController.js
@@ -13,40 +13,48 @@ export const getUserById = async (req, res) => {
         const user = await User.findById(id);
 
         if (!user) {
-            return res.status(404).send('User not found');
-        }
+            return res.status(404).send('User not found!');
+        };
 
-        res.status(200).send(user);
-    } catch(error) {
-        res.status(500).send('Something went wrong');
+        return res.status(200).send(user);
+    } catch (error) {
+        return res.status(500).send('Something went wrong. Please try again!');
     };
 };
 
 export const updateUser = async (req, res) => {
+    const { id } = req.params;
+
     try {
-        const { id } = req.params;
         const user = req.body;
 
-        const updatedUser = await User.findByIdAndUpdate(id, user);
+        if (id === req.user.id) {
+            const updatedUser = await User.findByIdAndUpdate(id, user);
 
-        res.status(200).send('User updated successfully');
+            if (updatedUser) {
+               return res.status(200).send('User updated successfully');
+            };
+        } else {
+            return res.status(403).send('Cannnot update user');
+        };
     } catch(error) {
-        res.status(500).send('Something went wrong');
+        return res.status(500).send('Something went wrong. Please try again!');
     };
 };
 
 export const deleteUser = async (req, res) => {
-    try {
-        const { id } = req.params;
-        
-        const deletedUser = await User.findByIdAndDelete(id);
+    const { id } = req.params;
 
-        if (deletedUser) {
-            res.status(200).send('User deleted successfully!');
+    try {
+        if (id === req.user.id) {
+            const deletedUser = await User.findByIdAndDelete(id);
+            if (deletedUser) {
+                return res.status(204).send();
+            }; 
         } else {
-            res.status(404).send('User not found!');
+            return res.status(404).send(`Could not delete user with id ${id}!`);
         };
-    } catch(error) {
-        res.status(500).send('Something went wrong!');
+    } catch (error) {
+        return res.status(500).send('Something went wrong. Please try again!');
     };
 };
